refactor: artistId 타입 Long으로 변경
change artistId type from Integer to Long across related classes

+++ frontend/BE/src/main/java/com/a404/duckonback/dto/ArtistDTO.java
@@ -11,7 +11,7 @@ import java.time.LocalDate;
 @AllArgsConstructor
 @Builder
 public class ArtistDTO {
-    private Integer artistId;
+    private Long artistId;
     private String nameEn;
     private String nameKr;
     private LocalDate debutDate;
